requirements.txt
fastapi
uvicorn
streamlit
requests
boto3
langchain-aws
langchain-core
python-dotenv
langchain-community
faiss-cpu